Separated styling in new file
